requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain>=0.0.470
sentence-transformers
huggingface-hub
faiss-cpu
PyPDF2
python-multipart
requests
numpy
PyYAML
langchain-community
pycryptodome
pycrypto